Clean up Makefile and bring it closer to conventions.
If the code should be compiled as C++, one should be using CXX as the
compiler in the makefile, and the flags passed to it should be CXXFLAGS.

It wasn't possible to use implicit rules because the source directory
and the output directory are different.

Use LDLIBS for the libraries, so external build systems can easily
inject libraries they need into the build. LDLIBS are setting with
target-specific variable values [1].

[1] https://www.gnu.org/software/make/manual/make.html#Target_002dspecific